+++ src/components/pages/HubView.vue
@@ -2,40 +2,22 @@
 // import components.
 import LastVisitedSkills from '../components/hub-components/LastVisitedSkills.vue';
 import Goals from '../components/hub-components/Goals.vue';
-import MarkAssessment from '../components/hub-components/MarkAssessment.vue';
-import HubStudentQuestionList from '../components/hub-components/HubStudentQuestionList.vue';
 import RecommendedSkillsGenerator from '../components/hub-components/RecommendedSkillsGenerator.vue';
-
 // Import store.
 import { useUserDetailsStore } from '../../stores/UserDetailsStore';
-import { useAssessmentsStore } from '../../stores/AssessmentsStore.js';
-import { useUsersStore } from '../../stores/UsersStore.js';
-import { useSkillsStore } from '../../stores/SkillsStore.js';
-import { useInstructorStudentsStore } from '../../stores/InstructorStudentsStore.js';
 
 export default {
     setup() {
         const userDetailsStore = useUserDetailsStore();
-        const assessmentsStore = useAssessmentsStore();
-        const usersStore = useUsersStore();
-        const skillsStore = useSkillsStore();
-        const instructorStudentsStore = useInstructorStudentsStore();
         // Run the GET request.
         userDetailsStore.getUserDetails();
         return {
-            userDetailsStore,
-            assessmentsStore,
-            usersStore,
-            skillsStore,
-            instructorStudentsStore
+            userDetailsStore
         };
     },
     data() {
         return {
             showWelcomeModal: false, // Modal to ask for tutorial preference
-            assessments: [],
-            questions: [],
-            studentIds: [],
             isTutorialComplete: false,
             showTutorialTip1: false,
             showTutorialTip2: false,
@@ -44,10 +26,8 @@ export default {
         };
     },
     components: {
-        MarkAssessment,
         LastVisitedSkills,
         Goals,
-        HubStudentQuestionList,
         RecommendedSkillsGenerator
     },
     computed: {
@@ -61,103 +41,8 @@ export default {
     },
     async created() {
         this.checkIfTutorialComplete();
-
-        if (this.userDetailsStore.role == 'instructor') {
-            await this.fetchAssessments();
-            await this.getStudentMCQuestions();
-        }
     },
     methods: {
-        async fetchAssessments() {
-            // Create the assessments array ---------------------------------
-            // Get unmarked assessments if there no assessment store before
-            await this.assessmentsStore.getAssessments();
-
-            // Get the instructor student list, if not yet loaded.
-            if (
-                this.instructorStudentsStore.instructorStudentsList.length == 0
-            ) {
-                await this.instructorStudentsStore.getInstructorStudentsList();
-            }
-
-            // Just get the students that this instructors teaches.
-            for (
-                let i = 0;
-                i < this.instructorStudentsStore.instructorStudentsList.length;
-                i++
-            ) {
-                if (
-                    this.userDetailsStore.userId ==
-                    this.instructorStudentsStore.instructorStudentsList[i]
-                        .instructor_id
-                ) {
-                    this.studentIds.push(
-                        this.instructorStudentsStore.instructorStudentsList[i]
-                            .student_id
-                    );
-                }
-            }
-            // Get the assessments for those students.
-            for (let i = 0; i < this.assessmentsStore.assessments.length; i++) {
-                for (let j = 0; j < this.studentIds.length; j++) {
-                    if (
-                        this.assessmentsStore.assessments[i].student_id ==
-                        this.studentIds[j]
-                    ) {
-                        this.assessments.push(
-                            this.assessmentsStore.assessments[i]
-                        );
-                    }
-                }
-            }
-
-            // Date.
-            for (let i = 0; i < this.assessments.length; i++) {
-                let date = new Date(this.assessments[i].date).toDateString();
-                this.assessments[i].date = date;
-            }
-
-            // Get users.
-            if (this.usersStore.users.length == 0) {
-                await this.usersStore.getUsers();
-            }
-            // Add the student name.
-            for (let i = 0; i < this.assessments.length; i++) {
-                for (let j = 0; j < this.usersStore.users.length; j++) {
-                    if (
-                        this.assessments[i].student_id ==
-                        this.usersStore.users[j].id
-                    ) {
-                        this.assessments[i].studentUsername =
-                            this.usersStore.users[j].username;
-                    }
-                }
-            }
-
-            // Get skills.
-            if (this.skillsStore.skillsList.length == 0) {
-                await this.skillsStore.getSkillsList();
-            }
-            // Add the skill name.
-            for (let i = 0; i < this.assessments.length; i++) {
-                for (let j = 0; j < this.skillsStore.skillsList.length; j++) {
-                    if (
-                        this.assessments[i].skill_id ==
-                        this.skillsStore.skillsList[j].id
-                    ) {
-                        this.assessments[i].skillName =
-                            this.skillsStore.skillsList[j].name;
-                    }
-                }
-            }
-        },
-        async getStudentMCQuestions() {
-            const result = await fetch(
-                '/questions/student-mc-questions/full-data-list'
-            );
-            const data = await result.json();
-            this.questions = data;
-        },
         // Tutorial
         async checkIfTutorialComplete() {
             try {
@@ -331,10 +216,6 @@ export default {
                         v-if="userDetailsStore.role == 'student'"
                         :userId="userDetailsStore.userId"
                     />
-                    <MarkAssessment
-                        v-if="userDetailsStore.role == 'instructor'"
-                        :assessments="assessments"
-                    />
                 </div>
             </div>
             <!-- Goals / Student Suggested Questions -->
@@ -364,11 +245,6 @@ export default {
                         </button>
                     </div>
                     <Goals />
-                    <!-- Student Added Questions List -->
-                    <HubStudentQuestionList
-                        v-if="userDetailsStore.role == 'instructor'"
-                        :questions="questions"
-                    />
                 </div>
             </div>
         </div>
